Some progress on Kafa-datasource JSR-233
