+++ datasource/opennms-kafka/pom.xml
@@ -119,6 +119,13 @@
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.codehaus.groovy</groupId>
+            <artifactId>groovy-all</artifactId>
+            <version>2.5.6</version>
+            <type>pom</type>
+            <scope>test</scope>
+        </dependency>
 
         <dependency>
             <groupId>org.apache.kafka</groupId>
