+++ datasource/opennms-kafka/src/main/java/org/opennms/oce/datasource/opennms/InventoryFromAlarm.java
@@ -1,0 +1,70 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2018 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.oce.datasource.opennms;
+
+import java.util.List;
+import java.util.Objects;
+
+import org.opennms.oce.datasource.opennms.proto.InventoryModelProtos;
+
+/**
+ * @author smith
+ *
+ */
+public class InventoryFromAlarm {
+
+    private final List<InventoryModelProtos.InventoryObject> inventory;
+
+    private final String managedObjectInstance;
+
+    private final String managedObjectType;
+
+    public InventoryFromAlarm(List<InventoryModelProtos.InventoryObject> inventory) {
+        this(inventory, null, null);
+    }
+
+    public InventoryFromAlarm(List<InventoryModelProtos.InventoryObject> inventory, String managedObjectInstance,
+            String managedObjectType) {
+        this.inventory = Objects.requireNonNull(inventory);
+        this.managedObjectInstance = managedObjectInstance;
+        this.managedObjectType = managedObjectType;
+    }
+
+    public List<InventoryModelProtos.InventoryObject> getInventory() {
+        return inventory;
+    }
+
+    public String getManagedObjectInstance() {
+        return managedObjectInstance;
+    }
+
+    public String getManagedObjectType() {
+        return managedObjectType;
+    }
+}
